created a working verision of weapon lore
Moved the weapon lore to a new file for the time being
